Some build speed improvements
Inside the container, mount/unmount a tmpfs to the build directory
for the build. Significantly speeds up build time.

+++ util/build_package.sh
@@ -3,12 +3,23 @@ build_package() {
 		return
 	fi
 
-	fetch_package
 	mkdir -p "$SYSROOT" "$PKGBUILDDIR"
+
+	if [ -z "$PYGOS_BUILD_CONTAINER" ]; then
+		fetch_package
+	else
+		mount -t tmpfs none "$PKGBUILDDIR"
+	fi
+
 	run_pkg_command "build"
 	run_pkg_command "deploy"
 	deploy_dev_cleanup
-	rm -rf "$PKGBUILDDIR"
+
+	if [ -z "$PYGOS_BUILD_CONTAINER" ]; then
+		rm -rf "$PKGBUILDDIR"
+	else
+		umount "$PKGBUILDDIR"
+	fi
 
 	touch "$PKGLOGDIR/${PKGNAME}.done"
 }
